missile placement rules
